Rename format 5 to format 3
